Fixed current behavior tests
