--- src/Workleap.DomainEventPropagation.Publishing.Tests/SampleDomainEvent.cs
@@ -1,14 +1,0 @@
-﻿namespace Workleap.DomainEventPropagation.Publishing.Tests;
-
-public class SampleDomainEvent : IDomainEvent
-{
-    public string? Message { get; set; }
-}
-
-public class SampleDomainEventHandler : IDomainEventHandler<SampleDomainEvent>
-{
-    public Task HandleDomainEventAsync(SampleDomainEvent domainEvent, CancellationToken cancellationToken)
-    {
-        return Task.CompletedTask;
-    }
-}
