+++ src/Workleap.DomainEventPropagation.Publishing.Tests/TracingBehaviorFixture.cs
@@ -1,5 +1,8 @@
-﻿using GSoft.Extensions.Xunit;
+﻿using Azure.Messaging.EventGrid;
+using FakeItEasy;
+using GSoft.Extensions.Xunit;
 using Microsoft.ApplicationInsights;
+using Microsoft.Extensions.Azure;
 using Microsoft.Extensions.DependencyInjection;
 using Workleap.DomainEventPropagation.Tests;
 
@@ -17,7 +20,15 @@ public class TracingBehaviorFixture : BaseUnitFixture
         // Application insights dependencies
         services.AddSingleton<TelemetryClient>();
 
+        services.AddOptions<EventPropagationPublisherOptions>()
+            .Configure(opt =>
+            {
+                opt.TopicEndpoint = "http://topic.entpoint";
+                opt.TopicAccessKey = "AccessKey";
+            });
+
         services.AddEventPropagationPublisher().AddApplicationInsights();
+        services.AddSingleton(A.Fake<IAzureClientFactory<EventGridPublisherClient>>());
 
         return services;
     }
